Fixed errors and a bug with dialogue code
Error with the Input.GetMouse would try to run without any code causing errors. Fixed by making sure the index needed to be set to 0 for it to call/work.

Needed to set index to -1 instead of leaving it at 0 with a bug discovered.

When you finished the dialogue the code would still try and call causing the TogglePlayer code to activate on and off ouside of dialogue. This was fixed because the input now requires the index to be at 0 or above but is now set to -1 at the start of the game and after each dialogue scene.